requirements.txt
pandas
numpy
python-dotenv
pybit
TA-Lib
requests
